@@ -49,7 +49,7 @@
     "transform": {
       ".(ts|tsx)": "ts-jest"
     },
-    "testEnvironment": "node",
+    "testEnvironment": "jsdom",
     "testRegex": "(/__tests__/.*|\\.(test|spec))\\.(ts|tsx|js)$",
     "moduleFileExtensions": [
       "ts",
@@ -62,14 +62,14 @@
     ],
     "coverageThreshold": {
       "global": {
-        "branches": 90,
-        "functions": 95,
-        "lines": 95,
-        "statements": 95
+        "branches": 65,
+        "functions": 75,
+        "lines": 75,
+        "statements": 75
       }
     },
     "collectCoverageFrom": [
-      "src/*.{ts,tsx}"
+      "src/**/*.{ts,tsx}"
     ]
   },
   "prettier": {
@@ -101,12 +101,11 @@
     "@semantic-release/git": "^7.0.7",
     "@semantic-release/github": "^5.2.9",
     "@semantic-release/npm": "^5.1.3",
-    "@types/enzyme": "^3.1.15",
-    "@types/enzyme-adapter-react-16": "^1.0.3",
     "@types/jest": "^23.3.2",
     "@types/jsdom": "^12.2.1",
     "@types/node": "^10.11.0",
     "@types/react": "^16.7.18",
+    "@types/react-dom": "^16.0.11",
     "@types/react-notification-system": "^0.2.39",
     "@types/sinon": "^7.0.3",
     "colors": "^1.3.2",
@@ -114,16 +113,16 @@
     "coveralls": "^3.0.2",
     "cross-env": "^5.2.0",
     "cz-conventional-changelog": "^2.1.0",
-    "enzyme": "^3.8.0",
-    "enzyme-adapter-react-16": "^1.7.1",
     "husky": "^1.0.1",
     "jest": "^23.6.0",
     "jest-config": "^23.6.0",
+    "jest-dom": "^3.0.0",
     "jsdom": "^13.1.0",
     "lint-staged": "^8.0.0",
     "lodash.camelcase": "^4.3.0",
     "prettier": "^1.14.3",
     "prompt": "^1.0.0",
+    "react-testing-library": "^5.4.4",
     "replace-in-file": "^3.4.2",
     "rimraf": "^2.6.2",
     "rollup": "^1.1.0",
@@ -144,15 +143,13 @@
     "typescript": "^3.0.3"
   },
   "dependencies": {
-    "react": "^16.7.0-alpha || ^16.8.0-alpha || ^16.8 || ^16.7",
-    "react-dom": "^16.7.0-alpha || ^16.8.0-alpha || ^16.8 || ^16.7",
-    "react-notification-system": ">0.2.7",
+    "react": "^16.8.0-alpha || ^16.8",
+    "react-dom": "^16.8.0-alpha || ^16.8",
     "redux": "^4.0.1"
   },
   "peerDependencies": {
-    "react-notification-system": ">0.2.7",
-    "react": "^16.7.0-alpha || ^16.8.0-alpha || ^16.8 || ^16.7",
-    "react-dom": "^16.7.0-alpha || ^16.8.0-alpha || ^16.8 || ^16.7",
+    "react": "^16.8.0-alpha || ^16.8",
+    "react-dom": "^16.8.0-alpha || ^16.8",
     "redux": "^4.0.1"
   }
 }
