@@ -16,7 +16,7 @@ export default {
     { file: pkg.module, format: 'es', sourcemap: true }
   ],
   // Indicate here external modules you don't wanna include in your bundle (i.e.: 'lodash')
-  external: ['react', 'react-dom', 'react-notification-system'],
+  external: ['react', 'react-dom', 'redux', 'prop-types'],
   watch: {
     include: 'src/**'
   },
